Move unit test from `test_fetch.py` to `test_au_si.py`
